JCR-1609 resolve TODO
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/trunk@779561 13f79535-47bb-0310-9956-ffa450edef68

+++ jackrabbit-spi-commons/src/main/java/org/apache/jackrabbit/spi/commons/query/sql2/Parser.java
@@ -933,10 +933,7 @@ public class Parser {
         }
         checkLiterals(false);
         
-        // TODO createValue(BigDecimal) is not yet implemented
-        // currentValue = valueFactory.createValue(bd);
-        currentValue = valueFactory.createValue(bd.doubleValue());
-        
+        currentValue = valueFactory.createValue(bd);
         currentTokenType = VALUE;
     }
